added state data

+++ db/migrate/20200828171655_create_us_states.rb
@@ -5,8 +5,8 @@ class CreateUsStates < ActiveRecord::Migration[6.0]
       t.string :abbrv
       t.string :nickname
       t.string :capital_city
-      t.integer :capital_lat
-      t.integer :capital_lng
+      t.float :capital_lat
+      t.float :capital_lng
       t.string :flag_img_url
 
       t.timestamps
